requirements.txt
absl-py==2.1.0
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
h5py==3.12.1
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.2.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
rich==13.9.4
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
